multisurveys-apis/pyproject.toml
[tool.poetry]
name = "multisurveys-apis"
version = "0.1.0"
description = ""
authors = ["Demurest <eric.apolonio@hotmail.com>"]
readme = "README.md"
packages = [
    { include = "core", from = "src" },
    { include = "object_api", from = "src" },
    { include = "lightcurve_api", from = "src" },
    { include = "magstat_api", from = "src" },
    { include = "classifier_api", from = "src" },
    { include = "probability_api", from = "src" },
    { include = "test_api", from = "src" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.19" }
returns = "^0.21.0"
psycopg2-binary = "^2.9.6"
prometheus-fastapi-instrumentator = "^6.1.0"
jinja2 = "^3.1.2"
httpx = "^0.24.1"
db-plugins = {git = "https://github.com/alercebroker/pipeline.git", subdirectory = "libs/db-plugins-multisurvey", branch = "multisurvey"}
djlint = "^1.35.2"
fastapi = "^0.115.12"
pydantic-settings = "^2.10.1"
pyyaml = "^6.0.2"
polars = "^1.31.0"
pyarrow = "^20.0.0"
pandas = "^2.3.1"
toolz = "^1.0.0"
astropy = "^7.1.0"
faker = "^37.6.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
testcontainers = "^4.12.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.14.1"

[tool.poetry.group.production.dependencies]
uvicorn = { extras = ["standard"], version = "^0.23.1" }

[tool.poetry.group.profiling.dependencies]
memray = "^1.8.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.12"

[build-system]
requires = ["poetry-core", "setuptools", "wheel", "Cython==0.29.36", "numpy"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.poetry.scripts]
all = "scripts.run_api:run"
object = "scripts.run_api:run_object"
lightcurve = "scripts.run_api:run_lightcurve"
magstat = "scripts.run_api:run_magstat"
classifier = "scripts.run_api:run_classifier"
probability = "scripts.run_api:run_probability"
dbp = "db_plugins.cli.manage:cli"
initdb = "scripts.initdb:initdb"
create_q3c_idx = "scripts.initdb:create_q3c_idx_cmd"
populate_db = "scripts.populate_database:main"
decode_oid = "scripts.decode_oid:main"

[pytest]
mock_use_standalone_module = true
